+++ example/sample.sh
@@ -1,1 +1,3 @@
+#!/usr/bin/sh
+
 echo "Hello from sample.sh"